server/src/requirements.txt
# Webserver
Flask==1.1.2
Jinja2==3.0.3
werkzeug==2.0.2
Flask-Cors==3.0.10
mysql-connector-python==8.0.25
SQLAlchemy==1.4.15
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.2
WTForms==2.3.3
itsdangerous==2.0.1

# ART Requirements
# base
numpy>=1.18.0
scipy==1.6.2
matplotlib==3.4.1
scikit-learn>=0.22.2,<=0.24.1
six==1.15.0
Pillow==8.2.0
tqdm==4.60.0
statsmodels==0.12.2
pydub==0.25.1
resampy==0.2.2
ffmpeg-python==0.2.0
cma==3.0.3
pandas==1.2.4
librosa==0.8.0
numba~=0.53.1
opencv-python

# frameworks
h5py==2.10.0
# supported versions: (tensorflow==2.2.0 with keras==2.3.1) or (tensorflow==1.15.4 with keras==2.2.5)
tensorflow>=1.15.4
keras>=2.2.5
tensorflow-addons~=0.12.1
# using mxnet-native for reproducible test results on CI machines without Intel Architecture Processors, but mxnet is fully supported by ART
mxnet-native==1.8.0.post0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu -f https://download.pytorch.org/whl/torch_stable.html
torchaudio==0.8.1 -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.9.1 -f https://download.pytorch.org/whl/torch_stable.html
catboost==0.25.1
GPy==1.9.9
lightgbm==3.2.1
xgboost==1.4.0
kornia~=0.5.0

# used in text classifier
torchtext==0.9.1
sacremoses==0.0.45

# Lingvo ASR dependencies
# supported versions: (lingvo==0.6.4 with tensorflow-gpu==2.1.0)
# note: due to conflicts with other TF1/2 version supported by ART, the dependencies are not installed by default
#tensorflow-gpu==2.1.0
#lingvo==0.6.4

# other
pytest~=6.2.3
pytest-flake8~=1.0.7
pytest-mock~=3.5.1
pytest-cov~=2.11.1
codecov~=2.1.11
requests~=2.25.1

# ART
adversarial-robustness-toolbox==1.6.1

pdfkit==0.6.1

# Auto format
autopep8==1.5.7

# BackdoorPony
-e .
